fix(app): update score context logic to match updated state management
